requirements.txt
flask
pymongo
flask-bcrypt
PyJWT
flask-pytest